refactor(recommended-questions): inline record acquisition into cron service
