+++ src/recommended-questions/recommended-questions-cron.service.ts
@@ -9,12 +9,20 @@ import { NamespacesService } from 'omniboxd/namespaces/namespaces.service';
 import { RecommendedQuestion } from 'omniboxd/recommended-questions/entities/recommended-question.entity';
 import { RecommendedQuestionsService } from 'omniboxd/recommended-questions/recommended-questions.service';
 import { TagService } from 'omniboxd/tag/tag.service';
-import { Repository } from 'typeorm';
+import { QueryFailedError, Repository } from 'typeorm';
 
-// Slightly below the cron interval so a pair scanned seconds after a tick
-// is not skipped at the next tick (which would double the effective cadence).
-const SCAN_FRESHNESS_MS = 9 * 60 * 1000;
+const SCAN_FRESHNESS_MS = 10 * 60 * 1000;
 const NAMESPACE_BATCH_SIZE = 100;
+
+const PG_UNIQUE_VIOLATION = '23505';
+
+function isUniqueViolation(err: unknown): boolean {
+  return (
+    err instanceof QueryFailedError &&
+    (err.driverError as { code?: string } | undefined)?.code ===
+      PG_UNIQUE_VIOLATION
+  );
+}
 
 function maxDate(...dates: (Date | undefined)[]): Date | undefined {
   let max: Date | undefined;
@@ -40,7 +48,7 @@ export class RecommendedQuestionsCronService {
     private readonly recommendedQuestionsService: RecommendedQuestionsService,
   ) {}
 
-  @Cron(CronExpression.EVERY_10_MINUTES, { waitForCompletion: true })
+  @Cron(CronExpression.EVERY_MINUTE, { waitForCompletion: true })
   @Span('RecommendedQuestionsCronService.scan')
   async scan(): Promise<void> {
     const counts = {
@@ -61,7 +69,7 @@ export class RecommendedQuestionsCronService {
         for (const userId of userIds) {
           counts.members++;
           try {
-            const generated = await this.checkRecommendQuestions(
+            const generated = await this.regenerateQuestions(
               namespaceId,
               userId,
             );
@@ -90,19 +98,29 @@ export class RecommendedQuestionsCronService {
     this.logger.log(`Recommended questions scan: ${JSON.stringify(counts)}`);
   }
 
-  async checkRecommendQuestions(
+  async regenerateQuestions(
     namespaceId: string,
     userId: string,
   ): Promise<boolean> {
     const now = new Date();
-    if (!(await this.shouldGenerate(namespaceId, userId, now))) {
+
+    const record = await this.acquireRecord(namespaceId, userId, now);
+    if (!record) {
       return false;
     }
-    const res =
-      await this.recommendedQuestionsService.generateRecommendedQuestions(
-        namespaceId,
-        userId,
-      );
+
+    const lastUpdated = await this.getLastUpdatedAt(namespaceId, userId);
+    if (!lastUpdated) {
+      return false;
+    }
+    if (record.generatedAt && record.generatedAt > lastUpdated) {
+      return false;
+    }
+
+    const res = await this.recommendedQuestionsService.generateQuestions(
+      namespaceId,
+      userId,
+    );
     await this.recommendedQuestionsRepository.update(
       { namespaceId, userId },
       {
@@ -117,39 +135,45 @@ export class RecommendedQuestionsCronService {
     return true;
   }
 
-  // Inserting the row on conflict-skip and compare-and-swapping scanned_at
-  // claim the pair, so concurrent instances never process it twice. If
-  // generation fails after the claim, the pair is simply rescanned once
-  // scanned_at ages out.
-  private async shouldGenerate(
+  private async acquireRecord(
     namespaceId: string,
     userId: string,
     now: Date,
-  ): Promise<boolean> {
-    const record = await this.recommendedQuestionsService.findOrInsert(
-      namespaceId,
-      userId,
-      now,
-    );
+  ): Promise<RecommendedQuestion | null> {
+    const record = await this.recommendedQuestionsRepository.findOne({
+      where: { namespaceId, userId },
+    });
 
-    if (now.getTime() - record.scannedAt.getTime() < SCAN_FRESHNESS_MS) {
-      return false;
+    if (!record) {
+      try {
+        await this.recommendedQuestionsRepository.insert({
+          namespaceId,
+          userId,
+          scannedAt: now,
+        });
+      } catch (err) {
+        if (isUniqueViolation(err)) {
+          return null;
+        }
+        throw err;
+      }
+      return await this.recommendedQuestionsRepository.findOneOrFail({
+        where: { namespaceId, userId },
+      });
     }
 
+    if (now.getTime() - record.scannedAt.getTime() < SCAN_FRESHNESS_MS) {
+      return null;
+    }
+    record.scannedAt = now;
     const result = await this.recommendedQuestionsRepository.update(
       { namespaceId, userId, scannedAt: record.scannedAt },
       { scannedAt: now },
     );
     if (result.affected !== 1) {
-      return false;
+      return null;
     }
-
-    const lastUpdated = await this.getLastUpdatedAt(namespaceId, userId);
-    if (!lastUpdated) {
-      return false;
-    }
-
-    return !record.generatedAt || record.generatedAt <= lastUpdated;
+    return record;
   }
 
   private async getLastUpdatedAt(
